Conflicts:
	h2o-docs/src/booklets/v2_2015/source/GLM_Vignette.tex
	h2o-r/H2O_Load.R
	h2o-r/h2o-package/R/frame.R
	h2o-r/h2o-package/R/glrm.R
	h2o-r/h2o-package/R/models.R
	h2o-r/h2o-package/R/parse.R
	h2o-r/tests/testdir_algos/glrm/runit_glrm_iris.R
	h2o-r/tests/testdir_algos/glrm/runit_glrm_missing.R
	h2o-r/tests/testdir_docexamples/runit_Rdoc_set_time_zone.R

+++ h2o-algos/src/test/java/hex/deeplearning/DeepLearningProstateTest.java
@@ -8,8 +8,10 @@ import org.junit.BeforeClass;
 import org.junit.Ignore;
 import org.junit.Test;
 import water.*;
+import water.exceptions.H2OModelBuilderIllegalArgumentException;
 import water.fvec.Frame;
 import water.fvec.NFSFileVec;
+import water.fvec.Vec;
 import water.parser.ParseDataset;
 import water.rapids.Env;
 import water.rapids.Exec;
@@ -45,13 +47,22 @@ public class DeepLearningProstateTest extends TestUtil {
       NFSFileVec vnfs = NFSFileVec.make(find_test_file(dataset));
       Frame vframe = ParseDataset.parse(Key.make(), vnfs._key);
 
-      Scope.enter();
       try {
         for (int resp : responses[i]) {
           boolean classification = !(i == 0 && resp == 2);
           if (classification && !frame.vec(resp).isEnum()) {
-            Scope.track(frame.replace(resp, frame.vec(resp).toEnum())._key);
-            DKV.put(frame._key, frame);
+            DKV.remove(frame._key);
+            String respname = frame.name(resp);
+            Vec r = frame.vec(respname).toEnum();
+            frame.remove(respname).remove();
+            frame.add(respname, r);
+            DKV.put(frame);
+
+            DKV.remove(vframe._key);
+            Vec vr = vframe.vec(respname).toEnum();
+            vframe.remove(respname).remove();
+            vframe.add(respname, vr);
+            DKV.put(vframe);
           }
           for (DeepLearningParameters.Loss loss : new DeepLearningParameters.Loss[]{
                   DeepLearningParameters.Loss.Automatic,
@@ -162,7 +173,6 @@ public class DeepLearningProstateTest extends TestUtil {
                                         if (fraction < rng.nextFloat()) continue;
 
                                         try {
-                                          Scope.enter();
                                           Log.info("**************************)");
                                           Log.info("Starting test #" + count);
                                           Log.info("**************************)");
@@ -213,7 +223,11 @@ public class DeepLearningProstateTest extends TestUtil {
                                             try {
                                               model1 = dl.trainModel().get();
                                               checkSums.add(model1.checksum());
-                                            } catch (Throwable t) {
+                                              testcount++;
+                                            } catch(Throwable t) {
+                                              model1 = DKV.getGet(p._model_id);
+                                              if (model1 != null)
+                                                Assert.assertTrue(model1._output._status == Job.JobState.FAILED);
                                               throw t;
                                             } finally {
                                               dl.remove();
@@ -281,12 +295,17 @@ public class DeepLearningProstateTest extends TestUtil {
                                             DeepLearning dl = new DeepLearning(p2);
                                             try {
                                               model2 = dl.trainModel().get();
-                                            } catch (Throwable t) {
+                                            } catch(Throwable t) {
+                                              model2 = DKV.getGet(p2._model_id);
+                                              if (model2 != null)
+                                                Assert.assertTrue(model2._output._status == Job.JobState.FAILED);
                                               throw t;
                                             } finally {
                                               dl.remove();
                                             }
                                           }
+                                          Assert.assertTrue(model1._output._status == Job.JobState.DONE);
+                                          Assert.assertTrue(model2._output._status == Job.JobState.DONE);
 
                                           assert(model1._parms != p2);
                                           assert(model1.model_info().get_params() != model2.model_info().get_params());
@@ -324,9 +343,10 @@ public class DeepLearningProstateTest extends TestUtil {
                                           assert(model2.model_info().get_params()._l2 == 1e-3);
 
                                           if (valid == null) valid = frame;
-                                          double threshold = 0;
+                                          double threshold;
                                           if (model2._output.isClassifier()) {
                                             Frame pred = null, pred2 = null;
+                                            Vec labels, predlabels, pred2labels;
                                             try {
                                               pred = model2.score(valid);
                                               // Build a POJO, validate same results
@@ -345,7 +365,9 @@ public class DeepLearningProstateTest extends TestUtil {
                                                 Assert.assertEquals(mm.cm().err(), error, 1e-15);
 
                                                 // check that the labels made with the default threshold are consistent with the CM that's reported by the AUC object
-                                                ConfusionMatrix cm = buildCM(valid.vecs()[resp].toEnum(), pred.vecs()[0].toEnum());
+                                                labels = valid.vec(resp);
+                                                predlabels = pred.vecs()[0];
+                                                ConfusionMatrix cm = buildCM(labels, predlabels);
                                                 Log.info("CM from pre-made labels:");
                                                 Log.info(cm.toASCII());
 //                                              Assert.assertEquals(cm.err(), error, 1e-4); //FIXME
@@ -363,7 +385,8 @@ public class DeepLearningProstateTest extends TestUtil {
                                                 } finally {
                                                   if (ev != null) ev.remove_and_unlock();
                                                 }
-                                                cm = buildCM(valid.vecs()[resp].toEnum(), pred2.vecs()[0].toEnum());
+                                                pred2labels = pred2.vecs()[0];
+                                                cm = buildCM(labels, pred2labels);
                                                 Log.info("CM from self-made labels:");
                                                 Log.info(cm.toASCII());
                                                 Assert.assertEquals(cm.err(), error, 1e-4); //AUC-given F1-optimal threshold might not reproduce AUC-given CM-error identically, but should match up to 1%
@@ -374,26 +397,28 @@ public class DeepLearningProstateTest extends TestUtil {
                                             }
                                           } //classifier
                                           else {
-                                            Frame pred = model2.score(valid);
-                                            // Build a POJO, validate same results
-                                            Assert.assertTrue(model2.testJavaScoring(frame,pred,1e-6));
-                                            pred.delete();
+                                            Frame pred = null;
+                                            try {
+                                              pred = model2.score(valid);
+                                              // Build a POJO, validate same results
+                                              Assert.assertTrue(model2.testJavaScoring(frame,pred,1e-6));
+                                            } finally {
+                                              if (pred!=null) pred.delete();
+                                            }
                                           }
                                           Log.info("Parameters combination " + count + ": PASS");
-                                          testcount++;
+                                        } catch (H2OModelBuilderIllegalArgumentException t) {
+                                          H2O.fail("should not get here");
                                         } catch (IllegalArgumentException t) {
-                                          continue; //next model
+                                          H2O.fail("should not get here");
+                                        } catch (RuntimeException t) {
+                                          Assert.assertTrue(t.getMessage().contains("unstable"));
                                         } catch (Throwable t) {
                                           t.printStackTrace();
                                           throw new RuntimeException(t);
                                         } finally {
-                                          if (model1 != null) {
-                                            model1.delete();
-                                          }
-                                          if (model2 != null) {
-                                            model2.delete();
-                                          }
-                                          Scope.exit();
+                                          if (model1 != null) model1.delete();
+                                          if (model2 != null) model2.delete();
                                         }
                                       }
                                     }
@@ -414,13 +439,15 @@ public class DeepLearningProstateTest extends TestUtil {
       } finally {
         frame.delete();
         vframe.delete();
-        Scope.exit();
       }
     }
-    Assert.assertTrue(checkSums.size() == testcount);
     Log.info("\n\n=============================================");
     Log.info("Tested " + testcount + " out of " + count + " parameter combinations.");
     Log.info("=============================================");
+    if (checkSums.size() != testcount) {
+      Log.info("Only found " + checkSums.size() + " unique checksums.");
+    }
+    Assert.assertTrue(checkSums.size() == testcount);
     }
 
     public static class Mid extends DeepLearningProstateTest {
